honksquad: refactor: extract organ logic to partial class and harden surgery validation
Move organ insertion/removal logic into SurgerySystem.Organs.cs partial class for
better separation. Add server-side validation for procedure selection (sender check,
range check, self-surgery block, hand-change break). Remove unused ClosesWound field
and InsertOrganEffect class. Grant surgery system access to Body/Organ components.
