+++ Content.Shared/Body/OrganComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.RussStation.Surgery.Systems; //HONK
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
@@ -8,7 +9,9 @@ namespace Content.Shared.Body;
 /// </summary>
 /// <seealso cref="BodySystem" />
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-[Access(typeof(BodySystem))]
+//HONK START - Surgery system needs to read organ Category
+[Access(typeof(BodySystem), typeof(SharedSurgerySystem))]
+//HONK END
 public sealed partial class OrganComponent : Component
 {
     /// <summary>
