--- Content.Shared/RussStation/Surgery/Effects/InsertOrganEffect.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.Serialization.Manager.Attributes;
-
-namespace Content.Shared.RussStation.Surgery.Effects;
-
-/// <summary>
-/// Surgery effect: inserts an organ held in the surgeon's off-hand into the patient.
-/// </summary>
-[DataDefinition]
-public sealed partial class InsertOrganEffect : ISurgeryEffect
-{
-}
