+++ Resources/Prototypes/@RussStation/Surgery/procedures.yml
@@ -35,7 +35,6 @@
       damage:
         types:
           Heat: 2
-      closesWound: true
 
 - type: surgeryProcedure
   id: TendWoundsBurn
@@ -73,7 +72,6 @@
       damage:
         types:
           Heat: 2
-      closesWound: true
 
 - type: surgeryProcedure
   id: OrganManipulation
